refactor: 코드 리뷰 반영 - 디바이스 상태 타입 통합 및 페이지 위치 정정 (#20)
- DeviceSection의 흩어진 상태를 DeviceInfoTypes 객체로 통합 (미사용 타입 활용)
- 라우팅 페이지(NotificationSettingPage, ProfileEditPage)를 components 하위에서 pages/setting 바로 아래로 이동
