+++ src/routes/AppRouter.tsx
@@ -6,8 +6,8 @@ import ModeEditPage from '@/pages/home/ModeEditPage';
 import Communication from '@/pages/communication/Communication';
 import LiveSound from '@/pages/liveSound/LiveSound';
 import Setting from '@/pages/setting/Setting';
-import NotificationSettingPage from '@/pages/setting/components/NotificationSettingPage';
-import ProfileEditPage from '@/pages/setting/components/profile/ProfileEditPage';
+import NotificationSettingPage from '@/pages/setting/NotificationSettingPage';
+import ProfileEditPage from '@/pages/setting/ProfileEditPage';
 import Login from '@/pages/login/Login';
 
 const AppRouter = () => {
